pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_coach"
version = "0.1.0"
description = "ИИ‑тренер для спортивного анализа"
requires-python = ">=3.11"
dependencies = [
    "streamlit==1.35.0",
    "openai==1.30.0",
    "pydantic==2.7.4",
    "python-dotenv==1.0.1",
    "pandas==2.2.2",
    "plotly==5.22.0"
]

[tool.setuptools.packages.find]
where = ["src"]
